feat: use connectedRender when possible
