+++ src/__tests__/edit-blending-price.test.tsx
@@ -1,5 +1,4 @@
 import {
-  render,
   screen,
   waitFor,
   waitForElementToBeRemoved,
@@ -7,8 +6,8 @@ import {
 import userEvent from "@testing-library/user-event";
 import { setupEthTesting } from "eth-testing";
 import { ethers } from "ethers";
-import AppProviders from "providers";
 import { contracts } from "rainbow-token-contracts";
+import { connectedRender } from "testing-utils";
 import App from "../App";
 
 describe("Edit blending price", () => {
@@ -73,7 +72,7 @@ describe("Edit blending price", () => {
         },
       });
 
-    render(<App />, { wrapper: AppProviders });
+    await connectedRender(<App />);
 
     await screen.findByText(/blending price: 1.0 ETH/i);
     await screen.findByText(/color: rgb\(123, 23, 124\)/i);
